Create Time loops and shapes
An example of getting Win32 a text based program to do time delayed animation.
I am unsure of when anyone would ever use this other than to just do it.
Please do not get into the habit of using system as code functions since they
will not do you any good in the future, but for now it is ok for the purpose of this
code. In addition you will see a lot of conversion from double to t_time errors.
Do not worry about this since the only threat to lost date would be within
your display which is not a problem for Win32.