Create VP MidTerm
This was my midterm for the first programming class at AI. Within this code
are functions, classes, class inheritance, an array, if statements, while loops,
switch statements, for loop time delays, and text and background color changes.
The code is not documented like it should be, but if you have any questions on the 
code or would like to improve upon it please feel free to contact me. It's a lot of
code, but for the final all this code should be at lease half the length by using multi-
dimensional arrays and mad crazy amounts of condensed and complex functions. Once again as
I always mention, try to stay away from system functions since they
wont do you any good later on, but for the purposes of Win32 test based
programming can simplify your code and your life.